requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
PyMuPDF==1.23.7
pandas==2.1.3
openpyxl==3.1.2
requests==2.31.0
python-multipart==0.0.6 
flask==2.3.3
reportlab==4.0.4
Werkzeug==2.3.7
supabase==2.15.0 